requirements.txt
absl-py==2.4.0
attrs==26.1.0
cffi==2.0.0
contourpy==1.3.3
cycler==0.12.1
flatbuffers==25.12.19
fonttools==4.62.1
jax==0.10.0
jaxlib==0.10.0
kiwisolver==1.5.0
matplotlib==3.10.8
mediapipe==0.10.13
ml_dtypes==0.5.4
MouseInfo==0.1.3
mysql-connector-python==9.6.0
numpy==2.4.4
opencv-contrib-python==4.13.0.92
opencv-python==4.13.0.92
opt_einsum==3.4.0
packaging==26.1
pillow==12.2.0
protobuf==4.25.9
PyAutoGUI==0.9.54
pycparser==3.0
PyGetWindow==0.0.9
PyMsgBox==2.0.1
pyobjc-core==12.1
pyobjc-framework-Cocoa==12.1
pyobjc-framework-Quartz==12.1
pyparsing==3.3.2
pyperclip==1.11.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
pytweening==1.2.0
rubicon-objc==0.5.3
scipy==1.17.1
six==1.17.0
sounddevice==0.5.5
